[mms] Due to issues with igbinary 1.1.1, disable the igbinary driver.
Unfortunately, there is a nasty bug with 1.1.1 that will cause PHP to
terminate without error or segfault on certain data.  I can reliably
trigger this when unserializing session data after logging into my test
Cyrus 2.4 account.  This is obviously unacceptable.

I have confirmed that the latest git release fixes this issue (from the
github repo).  But I don't feel the need to create a Horde-specific
release to remedy this at this time, so take the more conservative
approach of disabling entirely.